+++ HISTORY.rst
@@ -6,6 +6,7 @@
 **Bug fixes**
 
 - 2360_, [macOS]: can't compile on macOS < 10.13.  (patch by Ryan Schmidt)
+- 2254_, [Linux]: offline cpus raise NotImplementedError in cpu_freq() (patch by Shade Gladden)
 
 5.9.8
 =====