+++ HISTORY.rst
@@ -6,12 +6,15 @@
 **Enhancements**
 
 - 2366_, [Windows]: log debug message when using slower process APIs.
+- 2375_, [macOS]: provide arm64 wheels.  (patch by Matthieu Darbois)
 
 **Bug fixes**
 
-- 2360_, [macOS]: can't compile on macOS < 10.13.  (patch by Ryan Schmidt)
 - 2254_, [Linux]: offline cpus raise NotImplementedError in cpu_freq() (patch by Shade Gladden)
-- 2272_, Add pickle support to psutil Exceptions
+- 2272_: Add pickle support to psutil Exceptions.
+- 2360_, [macOS]: can't compile on macOS < 10.13.  (patch by Ryan Schmidt)
+- 2362_, [macOS]: can't compile on macOS 10.11.  (patch by Ryan Schmidt)
+- 2365_, [macOS]: can't compile on macOS < 10.9.  (patch by Ryan Schmidt)
 
 5.9.8
 =====